backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
psycopg==3.3.2
sqlparse==0.5.5
Pillow==11.1.0
